// frontend/mysite/src/interfaces/DB/DBdata.ts
export default interface DBdata {
  spaceID: number;
  aria: string;
  name: string;
  isCafe: boolean;
  address: string;
  station: string;
  wifi: number;
  isBattery: boolean;
  isMeetingRoom: boolean;
  price: number;
  dropInFeePerDay: number;
  dropInFeePerHour: number;
  startTime: string;
  endTime: string;
  isDropIn: boolean;
  jurisdiction: string;
  isEnhancedOptions: string;
  imageName: string;
  favorite: number;
  timeFromStation: number;
  MeetingRoomCapacity: number;
  options: string;
  density: number;
 }